# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carteira Financeira API",
        "description": "API para gerenciamento de carteira financeira",
        "contact": {
            "name": "Suporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notificacoes": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Lista todas as notificações",
                "description": "Retorna uma lista paginada de todas as notificações do usuário autenticado",
                "operationId": "4aa3bb6fa5b32347e2907d59a82406c6",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de notificações retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notificacao"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/MetaPaginacao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notificacoes/nao-lidas": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Lista notificações não lidas",
                "description": "Retorna uma lista paginada das notificações não lidas do usuário autenticado",
                "operationId": "98b7b54cf2760d43a06b0edd24c177e0",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de notificações não lidas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Notificacao"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/MetaPaginacao"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notificacoes/{id}/ler": {
            "post": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Marca notificação como lida",
                "description": "Marca uma notificação específica como lida",
                "operationId": "a0144555b92c995508e0208d359d7919",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação marcada como lida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Notificação marcada como lida"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado"
                    },
                    "404": {
                        "description": "Notificação não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notificacoes/ler-todas": {
            "post": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Marca todas as notificações como lidas",
                "description": "Marca todas as notificações do usuário autenticado como lidas",
                "operationId": "9a42897922b214334d0ec133161b9cf3",
                "responses": {
                    "200": {
                        "description": "Notificações marcadas como lidas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Todas as notificações foram marcadas como lidas"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notificacoes/nao-lidas/quantidade": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Conta notificações não lidas",
                "description": "Retorna o número total de notificações não lidas do usuário autenticado",
                "operationId": "0344e45d66d8359f1023356e7cad669e",
                "responses": {
                    "200": {
                        "description": "Quantidade de notificações não lidas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "quantidade": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Notificacao": {
                "title": "Notificação",
                "description": "Modelo de notificação conforme resposta da API",
                "properties": {
                    "id": {
                        "description": "ID interno da notificação",
                        "type": "integer",
                        "example": 101
                    },
                    "tipo": {
                        "description": "Tipo da notificação",
                        "type": "string",
                        "example": "info"
                    },
                    "titulo": {
                        "description": "Título da notificação",
                        "type": "string",
                        "example": "Saldo atualizado"
                    },
                    "mensagem": {
                        "description": "Conteúdo da notificação",
                        "type": "string",
                        "example": "Seu saldo foi atualizado com sucesso."
                    },
                    "dados": {
                        "description": "Dados adicionais da notificação",
                        "type": "string",
                        "example": {
                            "valor": 1500.75
                        }
                    },
                    "lida": {
                        "description": "Se a notificação foi lida",
                        "type": "boolean",
                        "example": false
                    },
                    "enviada": {
                        "description": "Se a notificação foi enviada",
                        "type": "boolean",
                        "example": true
                    },
                    "canal": {
                        "description": "Canal de envio da notificação",
                        "type": "string",
                        "example": "email"
                    },
                    "data_enviada": {
                        "description": "Data de envio da notificação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-10T16:00:00Z",
                        "nullable": true
                    },
                    "recurso_tipo": {
                        "description": "Tipo do recurso relacionado",
                        "type": "string",
                        "example": "conta"
                    },
                    "recurso_id": {
                        "description": "ID do recurso relacionado",
                        "type": "string",
                        "example": "c1a2b3c4-d5e6-7890-abcd-1234567890ef"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-10T15:55:00Z"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-10T16:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "MetaPaginacao": {
                "title": "Metadados de Paginação",
                "description": "Informações sobre a paginação dos resultados",
                "properties": {
                    "current_page": {
                        "description": "Página atual",
                        "type": "integer"
                    },
                    "per_page": {
                        "description": "Itens por página",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total de itens",
                        "type": "integer"
                    },
                    "total_pages": {
                        "description": "Total de páginas",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}